#refactor: consolidatde loan-related utility functions and enhance date formatting

+++ utils/global/formatDate.ts
@@ -1,7 +1,7 @@
 export function formatDate(dateStr: string) {
     try {
         return new Date(dateStr).toLocaleDateString("en-TZ", {
-            weekday: "long",
+            weekday: "short",
             year: "numeric",
             month: "long",
             day: "numeric",
